+++ internal/http/interface/context.hpp
@@ -1,0 +1,21 @@
+#ifndef CONTEXT_INTERFACE_HPP // NOTE: http/context.hpp と衝突しないようにしている
+#define CONTEXT_INTERFACE_HPP
+
+#include "http/request.hpp"
+#include "http/status.hpp"
+#include "task/io_task_manager.hpp"
+
+class IContext {
+public:
+    virtual ~IContext();
+    virtual const Request &getRequest() const = 0;
+    virtual void setRequest(const Request &request) = 0;
+    virtual void setHeader(const std::string &name, const std::string &value) = 0;
+    virtual void text(HttpStatusCode status, const std::string &body) = 0;
+    virtual void html(HttpStatusCode status, const std::string &body) = 0;
+    virtual void redirect(HttpStatusCode status, const std::string &location) = 0;
+    virtual IOTaskManager &getManager() const = 0;
+    virtual int getClientFd() const = 0;
+};
+
+#endif //CONTEXT_INTERFACE_HPP
