+++ internal/task/write_file.hpp
@@ -2,7 +2,7 @@
 #define WRITEFILE_HPP
 
 #include "callback_interface.hpp"
-#include "http/context.hpp"
+#include "http/interface/context.hpp"
 #include "io_task_manager.hpp"
 
 // NOLINTNEXTLINE(cppcoreguidelines-special-member-functions)
@@ -12,9 +12,9 @@ public:
     virtual Result<types::Unit, std::string> trigger() = 0;
 };
 
-class WriteFileCallback : public IWriteFileCallback {
+class CloseConnectionCallback : public IWriteFileCallback {
 public:
-    explicit WriteFileCallback(IContext *ctx);
+    explicit CloseConnectionCallback(IContext *ctx);
     Result<types::Unit, std::string> trigger();
 
 private:
@@ -23,13 +23,12 @@ private:
 
 class WriteFile : public IOTask {
 public:
-    WriteFile(IContext *ctx, const std::string &data_to_write, IWriteFileCallback *cb);
+    WriteFile(IOTaskManager &manager, int fd, const std::string &data_to_write, IWriteFileCallback *cb);
     ~WriteFile();
     virtual Result<IOTaskResult, std::string> execute();
 
 private:
     const std::string data_to_write_;
-    IContext *ctx_;
     IWriteFileCallback *cb_;
 };
